+++ api/v1/routes/auth.py
@@ -21,9 +21,8 @@ from api.db.database import get_db
 from api.utils.auth import authenticate_user, create_access_token,hash_password,get_user
 from api.utils.dependencies import get_current_admin, get_current_user
 from api.utils.json_response import JsonResponseDict
-from api.utils.config import SECRET_KEY, ALGORITHM
 from jose import JWTError
-import jwt
+from api.utils.auth import reset_password_request
 from api.v1.models.org import Organization
 
 from api.v1.models.product import Product
@@ -113,26 +112,7 @@ def read_admin_data(current_admin: Annotated[User, Depends(get_current_admin)]):
     return {"message": "Hello, admin!"}
 
 
-def reset_password_request(password, token, db: Session = Depends(get_db)):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-    email = payload.get("email", None)
-    if email is None:
-        raise credentials_exception
-    token_data = ResetPasswordTokenData(email=email)
-    user = db.query(User).filter(User.email == token_data.email).first()
-    if user is None:
-        raise credentials_exception
-    password_hashed = hash_password(password)
-    user.password = password_hashed
-    db.commit()
-    db.refresh(user)
-    db.close()
-    return True
+
 
 @auth.post("/reset-password")
 async def password_reset(request: ResetPasswordRequest, x_reset_token: Annotated[str | None, Header()] = None, db: Session = Depends(get_db)):
